+++ CHANGES.rst
@@ -114,6 +114,11 @@ New Features
 
 - Fix issue where catalog datasets could not be deleted from scatter viewers when aligned by WCS. [#4389]
 
+- Some modifications to spectral lines loader and table, including disabling
+  deletion of line name / rest wavelengths columns assigned by the loader, removing
+  the selection of line medium in loader, and adding line name to table by
+  default. [#4386]
+
 Mosviz
 ^^^^^^